refactor ContractAmendment components to improve form handling and validation
